Refactor core ensemble code
